+++ src/core/phases.rs
@@ -1,3 +1,4 @@
+use super::policy::phase_rank;
 use super::*;
 
 impl StateStore {